Consolidate blocking function Associate website blocking with a dedicated struct Refactor CLI commands for better clarity
