+++ src/cli/commands/start.rs
@@ -1,15 +1,29 @@
-// use crate::blocking::methods::block_duration;
-use crate::{blocking::websites::block_websites, os_backend};
+use duration_string::DurationString;
+
+use crate::blocking::websites::{self, Blockable};
+use crate::ui::spinners::show_interruptible_spinner_for_duration;
+use crate::ui::blocking_message::generate_blocking_message;
 
 pub fn cmd_start(start_args: &StartBlockArgs) -> Result<(), String> {
-    block_websites::block_websites_via_hosts_config_change(
-        &start_args.time_string,
-        start_args.task.as_ref(),
-    )
-    .map_err(|e| e.to_string())?;
+    let website_logic = websites::Websites::new();
 
-    let hosts_path = os_backend::get_hosts_path();
-    
+    let duration_to_wait = DurationString::from_string(start_args.time_string.to_string())
+        .map_err(|e| format!("Failed to parse duration: {}", e))?
+        .into();
+
+    let blocking_message =
+        generate_blocking_message(&start_args.time_string, start_args.task.as_ref());
+
+    website_logic
+        .block()
+        .map_err(|e| format!("Website blocking failed: {}", e))?;
+
+    show_interruptible_spinner_for_duration(&duration_to_wait, &blocking_message)
+        .map_err(|e| format!("Spinner Error: {}", e))?;
+
+    website_logic
+        .unblock()
+        .map_err(|e| format!("Website unblocking failed: {}", e))?;
 
     println!("\n  Unblocked websites ✅");
